refactor(card-browser): move browserColumnHeadings to fragment
Issue 17901
